fix: routing, forms, tests and removed test coverage

+++ src/Controller/TaskController.php
@@ -13,13 +13,13 @@ use Symfony\Component\HttpFoundation\Response;
 
 class TaskController extends AbstractController
 {
-    #[Route('/tasks', name: 'task_list')]
+    #[Route('/tasks', name: 'task_list', methods: ['GET'])]
     public function listAction(TaskRepository $taskRepository): Response
     {
         return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findAll()]);
     }
 
-    #[Route('/tasks/create', name: 'task_create')]
+    #[Route('/tasks/create', name: 'task_create', methods: ['GET', 'POST'])]
     public function createAction(Request $request, EntityManagerInterface $entityManager): Response
     {
         $task = new Task();
@@ -40,7 +40,7 @@ class TaskController extends AbstractController
         return $this->render('task/create.html.twig', ['form' => $form->createView()]);
     }
 
-    #[Route('/tasks/{id}/edit', name: 'task_edit')]
+    #[Route('/tasks/{id}/edit', name: 'task_edit', methods: ['GET', 'POST'])]
     public function editAction(Task $task, Request $request, EntityManagerInterface $entityManager): Response
     {
         $form = $this->createForm(TaskType::class, $task);
@@ -62,27 +62,35 @@ class TaskController extends AbstractController
         ]);
     }
 
-    #[Route('/tasks/{id}/toggle', name: 'task_toggle')]
-    public function toggleTaskAction(Task $task, EntityManagerInterface $entityManager): Response
+    #[Route('/tasks/{id}/toggle', name: 'task_toggle', methods: ['POST'])]
+    public function toggleTaskAction(Task $task, Request $request, EntityManagerInterface $entityManager): Response
     {
-        $task->toggle(!$task->isDone());
-        $entityManager->persist($task);
-        $entityManager->flush();
+        $token = $request->get('_token');
 
-        $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+        if ($this->isCsrfTokenValid('toggle-' . $task->getId(), $token)) {
+            $task->toggle(!$task->isDone());
+            $entityManager->persist($task);
+            $entityManager->flush();
+
+            $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+        }
 
         return $this->redirectToRoute('task_list');
     }
 
-    #[Route('/tasks/{id}/delete', name: 'task_delete')]
-    public function deleteTaskAction(Task $task, EntityManagerInterface $entityManager): Response
+    #[Route('/tasks/{id}/delete', name: 'task_delete', methods: ['POST'])]
+    public function deleteTaskAction(Task $task, Request $request, EntityManagerInterface $entityManager): Response
     {
         $this->denyAccessUnlessGranted('TASK_DELETE', $task, "Vous n'êtes pas autorisé à supprimer cette tâche.");
 
-        $entityManager->remove($task);
-        $entityManager->flush();
+        $token = $request->get('_token');
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+        if ($this->isCsrfTokenValid('delete-' . $task->getId(), $token)) {
+            $entityManager->remove($task);
+            $entityManager->flush();
+
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+        }
 
         return $this->redirectToRoute('task_list');
     }
